add cuda linear reg

+++ libs/dsi/basic_voxel.hpp
@@ -126,7 +126,7 @@ public://used in GQI
 public:// used in QSDR
     tipl::matrix<4,4> trans_to_mni;
     size_t template_id = 0;
-    tipl::transformation_matrix<double> qsdr_trans;
+    tipl::transformation_matrix<float> qsdr_trans;
     bool qsdr = false;
     tipl::vector<3,int> csf_pos1,csf_pos2,csf_pos3,csf_pos4;
     float R2;
